refactor(network): move media preview preparation into coordinator
